Refactor tests: rename sto_num to str_to_num
- Deleted strton_test.c as part of the cleanup process.
- In unicode_test.c, improved readability by formatting test results and ensuring consistency in test descriptions.
- In vec_test.c, replaced direct float comparisons with a comparison function to handle floating-point precision issues, enhancing the robustness of vector operation tests.
- Update filepath logic, cstr, array

+++ include/threadpool.h
@@ -23,7 +23,7 @@ STATIC_CHECK_POWER_OF_2(THREADLOCAL_QUEUE_SIZE);
 typedef struct threadpool ThreadPool;
 
 // Create and initialize a new threadpool with the given number of threads.
-ThreadPool* threadpool_create(int num_threads);
+ThreadPool* threadpool_create(size_t num_threads);
 
 // Add a new task to the threadpool. The function will be executed by one of
 // the threads in the pool.
